+++ runzi/tasks/oq_hazard/oq_disagg_task.py
@@ -21,7 +21,7 @@ from nzshm_model.psha_adapter.openquake import OpenquakeModelPshaAdapter
 from nzshm_model.psha_adapter.openquake.hazard_config import OpenquakeConfig
 from toshi_hazard_store.scripts.ths_disagg_import import store_disagg
 
-from runzi.arguments import EC2_JOB_DEFINITION, SubmissionArgs, TaskLanguage, TaskRuntimeArgs
+from runzi.arguments import EC2_JOB_DEFINITION, SubmissionArgs, TaskLanguage, TaskRuntimeArgs, serialize_arguments
 from runzi.automation.local_config import (
     API_URL,
     ECR_DIGEST,
@@ -145,8 +145,8 @@ class OQDisaggTask:
                 gmcm_logic_tree=json.dumps(gmcm_logic_tree.to_dict()),
                 openquake_config=json.dumps(openquake_config.to_dict()),
             ),
-            arguments=self.user_args.model_dump(
-                mode='json', exclude={'hazard_model', 'srm_logic_tree', 'gmcm_logic_tree', 'hazard_config'}
+            arguments=serialize_arguments(
+                self.user_args, exclude={'hazard_model', 'srm_logic_tree', 'gmcm_logic_tree', 'hazard_config'}
             ),
             # arguments={"a": 1},
             environment=environment,